Footnotes that end in anyting but a <p> (list, blockquote, codeblock) now have the backlink in a <p> which is appended to the end of the footnote. Fixes [1831600]
Also fixed multiparagraph footnotes so that the first paragraph is actually
wrapped in a <p> element. This results in one-liners also wrapped in <p> 
elements, but this is inline with PHP Markdown Extra's behavior. That means 
the test for `node.type == "text"` should never get a match, but I'm leaving 
it in for now.

Also added some tests, although they are not properly included in a test
directory for use with the testing framework. Seeing the framework currently
ignores extensions, we'll worry about that later.