Allow one or two spaces before a blockquote. Fixes [1899925]
That report actually askes for three spaces in imitation of markdown.pl.
However, the syntax rules overrule markdown.pl (even per J.G.) and the 
rules give no such allowance for any spaces. Such spaces have been a 
subject of debate on the list as well.

As it turns out, in python-markdown, allowing 3 spaces creates some 
problems for nested blockquotes (i.e.: a codeblock in a blockquote would 
need to be indented 5 spaces rather than 4.) Since we can't be breaking
that currently correct and expected behavior, I'll conceed and allow 2.

After all, markdown is supposed to be easier for humans to read and 
edit. One or two spaces are not so easy for the human eye to detect,
so we'll allow them. Three looks much more like 4, so I don't care
that we don't allow that many.