Fixed error handling to work when the logger is not being used.
Previously, when using as a module and not setting up your own logger, the
logger would fail and Markdown would choke on the unresolved problem. Now
exceptions or warnings are raised as appropriate (based on log level). However,
if the user has configured the logger, the errors are logged, not raised and
Markdown either continues or exits based on log level.