Rename docs/*.md => docs/*.txt
The documentation uses features of Python-Markdown that are not supported on
GitHub and it's better to get a source view of the docs anyway. For example,
that way comments and bug reports can reference a specific line of a file.

Of course, it makes sense for Github to render the README, so that is left
with the `.md` file extension.