Allow blockprocessor.run to return True or False.
This allows the run method to determine if a block is or is not a match in the
midst of parsing outside of the test method. The goal is to eliminate the
often redundant test method in the future. In the interim, it remains and
if the run method returns None, the existing behavior is maintained. Returning
None may generate a DepreciationWarning later. Note that test must still return
True to even get to the run method.