Add footnote content to as link title
When reading a reaosnable-sized document it's often inconvenient to jump
to the bottom of a document to read the footnote text. While this is
more or less unavoidable for large footnotes, short ones can
conveniently be displayed in a popup next to the footnote itself - this
is most easily achieved by setting the `title` attribute of the link.
This commit changes the `footnotes` extension such that the raw Markdown
source for the footnote content is added in such a capacity, truncated
to a specified width (160 characters by default) and with all whitespace
collapsed into single spaces.