Updated extension API docs to list output_format and serializer on the md instance.
I still need to document everything on that list and probably a few things missing from the list.

I also updated the warning about monkey patching to be a "Warning" admonition - now that we have that extension.

Probably should depreciate access to md_globals in the extension API -- as we don't actually use any anymore. That dates back to pre-OO code.