Allow inline attr_lists at end of header.
A header can be only one line - so unlike other block level elements, a
attr_list must be at th eend of the line. To disingush it from an inline
attr_list on the last child in a header, we must require at least one
space before th eblock-level attrt_list.

Always intended it to work that way. Not sure how I missed that before.
At least we're testing for it now.