Allow extensions to register serializers
Setting output_format must happen after extensions are loaded. Only in that
way can an extension register a serializer so that it will then be available
to be used with the output_format keyword. A test has been added to avoid
this regression from happening again in the future.

Fixes #238, partially reverses commit 41cc055 and provides a better fix for