Partial fix for #253.
This actually fixes the immediate problem. The method on the inlinePattern
returns the correct subtree. However, for some reason, the parent element
of that subtree is later stripped of all content (perhaps by the
inlineplaceholder stuff). So, for example, given this input:

    ***foo** bar*

The inlinepattern returns:

    <em><strong>foo</strong> bar</em>

However, later it is reduced to:

    <em></em>

Interestingly, this still works:

    ***foo***

Correctly becomes:

    <strong><em>foo</em></strong>

I'm guessing that the tail on the `em` tag in the first instance is tripping
things up some how.