Support extensions as modules.
Fixes #300. When a python module (.../__init__.py) is imported by `__import__`
and the `from_list` argument is a unicode string (as is returned by
rpartition), then an error is raised in Python 2.7. Force conversion to default
string type (byte in 2.7 and unicode in 3) to avoid the error.

For a full explaination, see issue
[300](https://github.com/waylan/Python-Markdown/issues/300).