Added `--extension_configs` option to the CLI.
The `--extension_configs` option must point to a YAML or JSON file.
The contents of the file must parse to a Python Dict which will be
passed to the `extension_configs` keyword of the `markdown.Markdown`
class.

Also added tests for all of the CLI option parsing options and updated
documentation.