Allow named extensions to specify the Class Name
If you were to import the class like this:

    from path.to.module import SomeExtensionClass

Then the named extension would be the string:

    "path.to.module:SomeExtensionClass"

This should simplify loading extensions from the command line or
template filters -- expecially when multiple extensions are defined
in a single python module.

The docs still need updating. I'm waiting to update the docs after
implementing #335 and #336 as that will require a major refactor of
that section of the docs anyway.