All extension tests now use python dot notation.
All named extensions now use python dot notation in the tests - including
all builtin extensions (eg: 'extra' => 'markdown.extensions.extra'). This
is in anticipation of #336.

Note there are a few tests (in the error tests) that will still need
updating, but not till we make the change as they will test for the
new error message.