Better nested STRONG EM support.
Addresses most of #253. Thanks to @facelessuser for the tests. Although
the tests currently failing may or may not be valid. It could be that
the failing tests will be removed with the expectation that when a
document author does that, they can see that it doesn't work and
change their document to something that works. But I need to go
through them before I'm sure.