Cleaned up STRNG EM tests.
Removed a bunch of weird ones (even some that passed). For the most part,
there is no definitive way for those to be parsed. So there is no point of
testing for them. In most of those situations, authors should be mixing
underscores and astericks so it is clear what is intended.