Simplify output_formats to html and xhtml.
We started with the numbers before HTML5 was a thing and we thought there
might be an XHTML2. Today, we know that all we have are HTML style tags and
XHTML style tags. Nothing else really matters in the real world.

Note that if '(x)html1' '(x)html4' or '(x)html5' are passed in, the number is
stripped/ignored. Users shouldn't need to change their code for this.