Use zip() for positional argument compatibility
The list returned by zip() is automatically truncated to the length of
the shortest argument list, which greatly simplifies the code that loops
through the list of positional arguments.

The expression `if len(args):` relies on the truthiness of `len(args)`;
using `if args:` does not change the behaviour of the conditional, but
is easier to understand at a glance (the length of args is irrelevant).