smart_emphasis keyword > legacy_em extension.
The smart_strong extension has been removed and its behavior is now the
default (smart em and smart strong are the default). The legacy_em
extension restores legacy behavior.

This completes the removal of keywords.  All parser behavior is now modified
by extensions, not by keywords on the Markdown class.