Inital attempt at an inline refactor.
This is very much incomplete and not fully tested. Likely still a
long way to go. At least it works with a some simple stuff. However,
tails are currenlty processed twice and various now-uneeded pieces
have not yet been removed.

Added a new treeprocessor which runs inlinepatterns without the need to use
placeholders or alter patterns. Note that any pattern that made use of
groups will need to be updated (reduce each group by 1) as a result.