Inline CSS class specification
The Rust documentation uses markdown that specifies CSS classes for fenced code, like ```{.classa .classb} which produces unbalanced fenced code currently. This should fix it so those classes are parsed and added to the resulting HTML, and the fenced code isn't made unbalanced.

Also removed that LANG_TAG substitution, since assuming the name of the class attribute in HTML to be "class" is pretty safe to make, plus not all CSS classes are only for specifying language highlighting.

No syntax errors in python2.7, python3.4 or pypy3