Nesting <pre> in <code> blocks instead
When I make a markdown message in the form of 4-space indent. 

        <4 spaces>test

The output is shown in the form of 

        <pre><code>test</code></pre>

This breaks some tagging in the wikimedia editor which I am using. (Quite an old version.). It interprets the nested code  tag as `<code>`

Simply inverting this should make it work again.