Fix lazy ordered list example in documentation
The documentation is not accurate when lazy_ol=False parameter is called.

The example shows the 'start' attribute to the first <li> tag. However, the attribute is in the <ol> tag:
>>> import markdown
>>> s = """
... 4. Apples
... 5. Oranges
... 6. Pears"""
>>> markdown.markdown(s, lazy_ol=False)
u'<ol start="4">\n<li>Apples</li>\n<li>Oranges</li>\n<li>Pears</li>\n</ol>'


The behaviour of the library is the correct one (https://developer.mozilla.org/fr/docs/Web/HTML/Element/ol), so the documentation need to be fixed, not the library.