Replace homegrown OrderedDict with purpose-built Registry.
All processors and patterns now get "registered" to a Registry.
Each item is given a name (string) and a priority. The name is for
later reference and the priority can be either an integer or float
and is used to sort. A Registry instance is a list-like iterable with
the items auto-sorted by priority. If two items have the same priority,
then they are listed in the order there were "registered". Registering
a new item with the same name as an already registered item replaces
the old item with the new item (however, the new item is sorted by
its priority). To remove an item, "deregister" it by name or index.

Fixes #418.