Increase flexibility of InlineProcessor
Allow additional analysis by passing in data buffer into handleMatch.  This allows us to give a simple regex for a start token, and then we can tokenize the content if we like inside handleMatch.  If we don't find what we are looking for, we can reject the match as if we found nothing. You must return your element and the bounds of the handled characters.