All Markdown instances are now 'md'.
Previously, instances of the Markdown class were represented as any one 
of 'md', 'md_instance', or 'markdown'. This inconsistancy made it 
diffcult when developing extensions, or just maintaining the existing 
code. Now, all instances are consistantly represented as 'md'.

The old attrbutes on class instances still exist, but raise a 
DeprecationWarning when accessed. Also on classes where the instance was 
optional, the attribute always exists now and is simply None if no 
instance was provided (previously the attribute wouldn't exist).