Account for code spans at start of line.
Previously we didn't test for this because we did not allow raw blocks
to have any indent. Now, that we allow up to 3 spaces of indent, we need
to confirm those 3 chars are actually whitespace. Tests added.

Note: the case where the code span is not on the first line is still
failing. I havn't worked out why but ran out of time today.