Prefer public packaging module to pkg_resources' private copy of it (#826)
Some distributions of `pkg_resources` remove the vendored copy of `packaging. Additionally, the vendored copy is not publicly documented and should not be relied on exclusively.  Regardless, if `pkg_resources` is installed at all, then we can be fairly certain one of `packaging` or `pkg_resources._vendor.packaging` will be available. That being the case, there is no need to list `packaging` as a dependency.

Fixes #825.