Remove setuptools runtime dependency
Adding setuptools to install_requires does not play nice with pip-tools.
It leaves a very large warning at the bottom of requirements.txt files:

    # WARNING: The following packages were not pinned, but pip requires them to be
    # pinned when the requirements file includes hashes. Consider using the --allow-unsafe flag.
    # setuptools==41.0.1        # via markdown

This becomes an issue when installing packages with --require-hashes as
now setuptools is not pinned. This results in the error:

    In --require-hashes mode, all requirements must have their versions pinned with ==. These do not:
        setuptools>=36 from ... (from markdown==3.1.1->-r .../requirements.txt (line 30))

It is only being used to generate a version string. But this is very
simple to implement without setuptools or packaging.