Remove tox.ini from MANIFEST
Tox has already been removed from `make test` and is no longer reccomended
for testing in the contributing guide. Additionally, the extra tests it runs
(checking the spelling of documentation, etc) are of no value to end users.
As things are currently set us, tox is only used on the Travis CI server.
Given the above, there is no reason to include it in the release tarball.