Avoid importing packaging or pkg_resources for version validation (#948)
Importing the `pkg_resources` module has high memory and startup time
cost. A recent change in 102e01c already avoided it for loading extensions,
but it's still used for validating that __version__ is correctly formatted.

It is possible to avoid it by installing the `packaging` package, but
that adds a dependency for something quite trivial.

Instead, remove the validation and add tests which check the output is
as expected.

Since `setuptools` is no longer required at runtime, remove it from
`install_required`.